Corrected many issues with Godot 4.4
The plugin and example implementation did not work with Godot 4.4.  Updated all files to make it work.  Also added in the use of class files attached to states so the signal calling back to a main node is not needed.  Will need to further test that this works with stacked fsm implementation.